backend/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
channels==2.4.0
django-filter==2.2.0
djangochannelsrestframework==0.0.5
drf-yasg==1.17.1
pytest==5.4.0
pytest-django==3.9.0
pytest-asyncio==0.11.0
channels_redis==2.4.2
Pillow==10.1.0
openpyxl==3.1.2
bs4==0.0.1
requests
easy-thumbnails==2.8.5
django-image-cropping==1.4.0
django-cors-headers==3.11.0
djangorestframework-jwt==1.11.0
